breaking down into multiple components
