+++ CHANGELOG.md
@@ -7,11 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+-   Moved the API tokens management to a new page supporting multiple token per user (#198)
+
 ## [0.41.0] - 2023-05-11
-
-### Changed
-
--   Serialize the performance graph regarding the identifier only (#194)
 
 ### Added
 
@@ -20,10 +20,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
--   Moved the API tokens management to a new page supporting multiple token per user (#198)
 -   Upgrade node version from 16.13.0 to 18.16.0 (#187)
 -   New icon for list ordering in tables (#191)
 -   CP cancel alert message (#192)
+-   Serialize the performance graph regarding the identifier only (#194)
 
 ### Fixed
